+++ frontend/src/metabase/visualizations/components/legend/LegendCaption/LegendCaption.tsx
@@ -8,9 +8,10 @@ import DashboardS from "metabase/css/dashboard.module.css";
 import { useTranslateContent } from "metabase/i18n/hooks";
 import EmbedFrameS from "metabase/public/components/EmbedFrame/EmbedFrame.module.css";
 import type { IconProps } from "metabase/ui";
-import { Tooltip } from "metabase/ui";
+import { Icon, Menu, Tooltip } from "metabase/ui";
 
 import LegendActions from "../LegendActions";
+
 import {
   LegendCaptionRoot,
   LegendDescriptionIcon,
@@ -40,6 +41,7 @@ interface LegendCaptionProps {
   actionButtons?: React.ReactNode;
   hasInfoTooltip?: boolean;
   onSelectTitle?: () => void;
+  titleMenuItems?: React.ReactNode;
   width?: number;
 }
 
@@ -53,6 +55,7 @@ export const LegendCaption = ({
   hasInfoTooltip = true,
   onSelectTitle,
   width,
+  titleMenuItems,
 }: LegendCaptionProps) => {
   /*
    * Optimization: lazy computing the href on title focus & mouseenter only.
@@ -78,32 +81,52 @@ export const LegendCaption = ({
 
   const tc = useTranslateContent();
   const translatedTitle = tc(title);
-  const translatedDescription = tc(description);
+
+  const titleElement = (
+    <LegendLabel
+      className={cx(
+        DashboardS.fullscreenNormalText,
+        DashboardS.fullscreenNightText,
+        EmbedFrameS.fullscreenNightText,
+      )}
+      href={href}
+      onClick={onSelectTitle}
+      onFocus={handleFocus}
+      onMouseEnter={handleMouseEnter}
+    >
+      <Ellipsified data-testid="legend-caption-title">
+        {translatedTitle}
+      </Ellipsified>
+      {titleMenuItems && (
+        <Icon
+          style={{ flexShrink: 0, marginRight: 10 }}
+          name="chevrondown"
+          size={10}
+          className={CS.ml1}
+        />
+      )}
+    </LegendLabel>
+  );
 
   return (
     <LegendCaptionRoot className={className} data-testid="legend-caption">
       {icon && <LegendLabelIcon {...icon} />}
-      <LegendLabel
-        className={cx(
-          DashboardS.fullscreenNormalText,
-          DashboardS.fullscreenNightText,
-          EmbedFrameS.fullscreenNightText,
-        )}
-        href={href}
-        onClick={onSelectTitle}
-        onFocus={handleFocus}
-        onMouseEnter={handleMouseEnter}
-      >
-        <Ellipsified data-testid="legend-caption-title">
-          {translatedTitle}
-        </Ellipsified>
-      </LegendLabel>
+      {titleMenuItems ? (
+        <Menu>
+          <Menu.Target>{titleElement}</Menu.Target>
+          <Menu.Dropdown data-testid="legend-caption-menu">
+            {titleMenuItems}
+          </Menu.Dropdown>
+        </Menu>
+      ) : (
+        titleElement
+      )}
       <LegendRightContent>
         {hasInfoTooltip && description && !shouldHideDescription(width) && (
           <Tooltip
             label={
               <Markdown dark disallowHeading unstyleLinks lineClamp={8}>
-                {translatedDescription}
+                {description}
               </Markdown>
             }
             maw="22em"
